watch lebel version, belnder
git-svn-id: https://cgru.svn.sourceforge.net/svnroot/cgru@1917 0f86f178-d61e-0410-9d12-f6255c3815b8

+++ afanasy/trunk/python/parsers/blender.py
@@ -44,7 +44,7 @@ class blender(parser.parser):
                      ok = False
                   if ok:
                      if part1 > 0:
-                        self.percentframe = 100 * part0 / part1
+                        self.percentframe = int( 100 * part0 / part1)
                         need_calc = True
 
             if need_calc: self.calculate()
